Move `sightglass_analyze::summarize` to its own module
